WIP: Refactor common stuff, package.json scripts, etc.
