+++ lib/spectacle/actions.js
@@ -7,7 +7,7 @@ const webpack = require('webpack');
 const WebpackDevServer = require('webpack-dev-server');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-const baseCfg = require('../../config/webpack.config');
+const baseCfg = require('../../config/webpack.config.cli');
 
 // Injection replacement patterns.
 const MDX_RE = /\/\/ SPECTACLE_CLI_MDX_START[\s\S]*?\/\/ SPECTACLE_CLI_MDX_END/gm;
@@ -57,7 +57,7 @@ const webpackConfig = async ({
   // Use a simplified "markdown only" template for `.md` files.
   const isMd = /\.md$/.test(srcFilePath);
   const isMdx = /\.mdx$/.test(srcFilePath);
-  const tmplDir = `templates/${isMd ? 'md' : 'mdx'}-slides`;
+  const tmplDir = `../templates/${isMd ? 'md' : 'mdx'}-slides`;
 
   // Webpack compiler + configuration.
   const entry = path.resolve(__dirname, `${tmplDir}/index.js`);
