+++ package.json
@@ -15,16 +15,17 @@
     "clean:dist": "rimraf \"examples/*/dist\"",
     "clean:boilerplate": "rimraf .examples-boilerplate",
     "clean": "yarn clean:dist && yarn clean:boilerplate",
-    "build-boilerplate": "echo 'TODO(bp): Implement boilerplate build'",
-    "build-examples-cli-md": "cd examples/cli-md && node ../../bin/spectacle/cli.js --title \"Spectacle CLI + MD Example\" --src slides.md -a build",
-    "build-examples-cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Custom Theme Example\" -t theme.js -f true --action build --output dist",
-    "build-examples-cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Spectacle Slide Example\" -q template.js --action build",
-    "build-examples-loader-mdx": "webpack --config examples/loader-mdx/webpack.config.js",
-    "build-examples": "yarn build-examples-cli-md && yarn build-examples-cli-mdx && yarn build-examples-loader-mdx",
+    "build:boilerplate": "echo 'TODO(bp): Implement boilerplate build'",
+    "build:examples-cli-md": "cd examples/cli-md && node ../../bin/spectacle/cli.js --title \"Spectacle CLI + MD Example\" --src slides.md -a build",
+    "build:examples-cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Custom Theme Example\" -t theme.js -f true --action build --output dist",
+    "build:examples-cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Spectacle Slide Example\" -q template.js --action build",
+    "build:examples-loader-mdx": "webpack --config examples/loader-mdx/webpack.config.js",
+    "build:examples": "builder concurrent --buffer build:examples-cli-md build:examples-cli-mdx build:examples-cli-mdx-babel build:examples-loader-mdx",
     "start:cli-md": "cd examples/cli-md && node ../../bin/spectacle/cli.js --title \"Spectacle CLI + MD Example\" --src slides.md --port 3100",
     "start:cli-mdx": "cd examples/cli-mdx && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Custom Theme Example\" -t theme.js -f true -p 3000",
     "start:cli-mdx-babel": "cd examples/cli-mdx-babel && node ../../bin/spectacle/cli.js -l \"Spectacle CLI + MDX + Babel + Spectacle Slide Example\" -q template.js -p 3001",
     "start:loader-mdx": "webpack-dev-server --port=4000 --hot --config examples/loader-mdx/webpack.config.js",
+    "start:examples": "builder concurrent --log-level=info start:cli-md start:cli-mdx start:cli-mdx-babel start:loader-mdx",
     "start": "yarn start:cli-mdx",
     "lint": "eslint .",
     "check": "yarn lint",
@@ -38,6 +39,7 @@
   "devDependencies": {
     "babel-eslint": "^10.0.2",
     "babel-plugin-codegen": "^3.1.0",
+    "builder": "^5.0.0",
     "eslint": "^6.7.2",
     "eslint-config-formidable": "^4.0.0",
     "eslint-config-prettier": "^6.7.0",
